Move the snippet editor wrapper
It doesn't need to be tested.
